pyproject.toml
[tool.poetry]
name = "fastapi-mongodb-base"
version = "0.1.0"
description = ""
authors = ["Kengkkun <waengchaikkun@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_mongodb_base"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
pydantic-settings = "^2.0.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
asyncio = "^3.4.3"
pytest-asyncio = "^0.21.1"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "^2.1.1"}
umongo = "^3.1.0"
motor = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


